refactor(portfolio): introduce ChatSettings context and split oversized hook (#665)
Replace 39-prop drilling through ChatSettings → ChatSettingsForm with
React Context. Split the monolithic useChatSettings hook (280 lines)
into three focused hooks: useLocalStorageSettings, useModelFetching,
and useSettingsHandlers, keeping useChatSettings as a thin facade.
