+++ projects/portfolio/src/client/components/chat/chat-settings/chat-settings-form.tsx
@@ -1,73 +1,28 @@
 import { ToggleInput } from '#/client/components/input/toggle-input'
-import type { Settings } from '#/client/storage/remote-storage-settings'
-import { type ChangeEvent } from 'react'
+import { useChatSettingsContext } from './chat-settings-context'
 import { ModelSelector } from './model-selector'
 import { AutoModelToggle } from './settings/auto-model-toggle'
 import { ReasoningEffort } from './settings/reasoning-effort'
 import { TemperatureSlider } from './settings/temperature-slider'
 import { TextInput } from './settings/text-input'
 
-interface Props {
-  settings: Settings
-  fetchedModels: string[]
-  isLoadingModels: boolean
-  fetchError: string | null
-  temperature: number
-  temperatureEnabled: boolean
-  autoModel: boolean
-  fakeMode: boolean
-  markdownPreview: boolean
-  streamMode: boolean
-  interactiveMode: boolean
-  reasoningEffort: 'none' | 'minimal' | 'low' | 'medium' | 'high' | 'xhigh'
-  reasoningEffortEnabled: boolean
-  onChangeAutoModel: (event: ChangeEvent<HTMLSelectElement>) => void
-  onChangeManualModel: (event: ChangeEvent<HTMLInputElement>) => void
-  onChangeBaseURL: (event: ChangeEvent<HTMLInputElement>) => void
-  onChangeApiKey: (event: ChangeEvent<HTMLInputElement>) => void
-  onChangeMcpServerURLs: (event: ChangeEvent<HTMLInputElement>) => void
-  onChangeTemperature: (event: ChangeEvent<HTMLInputElement>) => void
-  onChangeMaxTokens: (event: ChangeEvent<HTMLInputElement>) => void
-  onToggleTemperature: () => void
-  onToggleAutoModel: () => void
-  onToggleFakeMode: () => void
-  onToggleMarkdownPreview: () => void
-  onToggleStreamMode: () => void
-  onToggleInteractiveMode: () => void
-  onToggleReasoningEffort: () => void
-  onChangeReasoningEffort: (event: ChangeEvent<HTMLSelectElement>) => void
-}
+export function ChatSettingsForm() {
+  const {
+    settings,
+    fakeMode,
+    markdownPreview,
+    streamMode,
+    interactiveMode,
+    handleChangeBaseURL,
+    handleChangeApiKey,
+    handleChangeMcpServerURLs,
+    handleChangeMaxTokens,
+    handleToggleFakeMode,
+    handleToggleMarkdownPreview,
+    handleToggleStreamMode,
+    handleToggleInteractiveMode,
+  } = useChatSettingsContext()
 
-export function ChatSettingsForm({
-  settings,
-  fetchedModels,
-  isLoadingModels,
-  fetchError,
-  temperature,
-  temperatureEnabled,
-  autoModel,
-  fakeMode,
-  markdownPreview,
-  streamMode,
-  interactiveMode,
-  reasoningEffort,
-  reasoningEffortEnabled,
-  onChangeAutoModel,
-  onChangeManualModel,
-  onChangeBaseURL,
-  onChangeApiKey,
-  onChangeMcpServerURLs,
-  onChangeTemperature,
-  onChangeMaxTokens,
-  onToggleTemperature,
-  onToggleAutoModel,
-  onToggleFakeMode,
-  onToggleMarkdownPreview,
-  onToggleStreamMode,
-  onToggleInteractiveMode,
-  onToggleReasoningEffort,
-  onChangeReasoningEffort,
-}: Props) {
   return (
     <div className='flex flex-col gap-5'>
       {/* Model Section */}
@@ -81,20 +36,11 @@ export function ChatSettingsForm({
             >
               Model
             </label>
-            <ModelSelector
-              model={settings.model}
-              autoModel={autoModel}
-              fakeMode={fakeMode}
-              fetchedModels={fetchedModels}
-              isLoadingModels={isLoadingModels}
-              fetchError={fetchError}
-              onChangeAutoModel={onChangeAutoModel}
-              onChangeManualModel={onChangeManualModel}
-            />
+            <ModelSelector />
           </div>
 
           {/* Auto Model Toggle */}
-          <AutoModelToggle autoModel={autoModel} onClick={onToggleAutoModel} />
+          <AutoModelToggle />
         </div>
       </section>
 
@@ -108,7 +54,7 @@ export function ChatSettingsForm({
             defaultValue={settings.baseURL || 'https://api.openai.com/v1'}
             placeholder='https://api.openai.com/v1'
             disabled={fakeMode}
-            onChange={onChangeBaseURL}
+            onChange={handleChangeBaseURL}
           />
 
           <TextInput
@@ -118,7 +64,7 @@ export function ChatSettingsForm({
             defaultValue={settings.apiKey}
             placeholder='Enter your API key'
             disabled={fakeMode}
-            onChange={onChangeApiKey}
+            onChange={handleChangeApiKey}
           />
 
           <TextInput
@@ -128,7 +74,7 @@ export function ChatSettingsForm({
             defaultValue={settings.mcpServerURLs || ''}
             placeholder='http://localhost:3001, http://localhost:3002'
             disabled={fakeMode}
-            onChange={onChangeMcpServerURLs}
+            onChange={handleChangeMcpServerURLs}
           />
         </div>
       </section>
@@ -137,12 +83,7 @@ export function ChatSettingsForm({
       <section className='space-y-3'>
         <h3 className='text-sm font-medium text-gray-500 uppercase dark:text-gray-400'>Parameters</h3>
         <div className='space-y-4'>
-          <TemperatureSlider
-            temperature={temperature}
-            temperatureEnabled={temperatureEnabled}
-            onChange={onChangeTemperature}
-            onToggle={onToggleTemperature}
-          />
+          <TemperatureSlider />
 
           <TextInput
             name='maxTokens'
@@ -152,15 +93,10 @@ export function ChatSettingsForm({
             max={4096}
             defaultValue={settings.maxTokens?.toString()}
             placeholder='Max tokens'
-            onChange={onChangeMaxTokens}
+            onChange={handleChangeMaxTokens}
           />
 
-          <ReasoningEffort
-            reasoningEffort={reasoningEffort}
-            reasoningEffortEnabled={reasoningEffortEnabled}
-            onChange={onChangeReasoningEffort}
-            onToggle={onToggleReasoningEffort}
-          />
+          <ReasoningEffort />
         </div>
       </section>
 
@@ -172,19 +108,19 @@ export function ChatSettingsForm({
             label='Markdown Preview'
             labelClassName='text-sm font-medium text-gray-700 dark:text-gray-300'
             value={markdownPreview}
-            onClick={onToggleMarkdownPreview}
+            onClick={handleToggleMarkdownPreview}
           />
           <ToggleInput
             label='Stream Mode'
             labelClassName='text-sm font-medium text-gray-700 dark:text-gray-300'
             value={streamMode}
-            onClick={onToggleStreamMode}
+            onClick={handleToggleStreamMode}
           />
           <ToggleInput
             label='Interactive Mode'
             labelClassName='text-sm font-medium text-gray-700 dark:text-gray-300'
             value={interactiveMode}
-            onClick={onToggleInteractiveMode}
+            onClick={handleToggleInteractiveMode}
           />
         </div>
       </section>
@@ -196,7 +132,7 @@ export function ChatSettingsForm({
           label='Fake Mode'
           labelClassName='text-sm font-medium text-gray-700 dark:text-gray-300'
           value={fakeMode}
-          onClick={onToggleFakeMode}
+          onClick={handleToggleFakeMode}
         />
       </section>
