+++ projects/portfolio/src/client/components/chat/chat-settings/index.ts
@@ -2,3 +2,4 @@ export { ChatSettings } from './chat-settings'
 export { ChatSettingsPanel } from './chat-settings-panel'
 export { ChatSettingsForm } from './chat-settings-form'
 export { ModelSelector } from './model-selector'
+export { ChatSettingsProvider, useChatSettingsContext } from './chat-settings-context'
