+++ projects/portfolio/src/client/components/chat/chat-settings/model-selector.tsx
@@ -1,26 +1,17 @@
-import { type ChangeEvent } from 'react'
+import { useChatSettingsContext } from './chat-settings-context'
 
-interface Props {
-  model: string
-  autoModel: boolean
-  fakeMode: boolean
-  fetchedModels: string[]
-  isLoadingModels: boolean
-  fetchError: string | null
-  onChangeAutoModel: (event: ChangeEvent<HTMLSelectElement>) => void
-  onChangeManualModel: (event: ChangeEvent<HTMLInputElement>) => void
-}
+export function ModelSelector() {
+  const {
+    settings,
+    autoModel,
+    fakeMode,
+    fetchedModels,
+    isLoadingModels,
+    fetchError,
+    handleChangeAutoModel,
+    handleChangeManualModel,
+  } = useChatSettingsContext()
 
-export function ModelSelector({
-  model,
-  autoModel,
-  fakeMode,
-  fetchedModels,
-  isLoadingModels,
-  fetchError,
-  onChangeAutoModel,
-  onChangeManualModel,
-}: Props) {
   if (autoModel) {
     return (
       <div className='space-y-1'>
@@ -31,9 +22,9 @@ export function ModelSelector({
               ? 'cursor-not-allowed border-gray-200 bg-gray-50 text-gray-400'
               : 'border-gray-300 bg-white text-gray-900 hover:border-gray-400 focus:border-blue-500 focus:ring-2 focus:ring-blue-200 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:focus:ring-blue-800'
           }`}
-          onChange={onChangeAutoModel}
+          onChange={handleChangeAutoModel}
           disabled={fakeMode || isLoadingModels || fetchedModels.length === 0}
-          value={model}
+          value={settings.model}
         >
           {isLoadingModels ? (
             <option>Loading...</option>
@@ -61,9 +52,9 @@ export function ModelSelector({
   return (
     <input
       name='model'
-      defaultValue={model}
+      defaultValue={settings.model}
       disabled={fakeMode}
-      onChange={onChangeManualModel}
+      onChange={handleChangeManualModel}
       placeholder='Enter model name'
       className={`w-full rounded-md border px-3 py-2 text-sm outline-none transition-all ${
         fakeMode
