+++ projects/portfolio/src/client/components/chat/chat-settings/settings/auto-model-toggle.tsx
@@ -1,17 +1,15 @@
 import { ToggleInput } from '#/client/components/input/toggle-input'
+import { useChatSettingsContext } from '../chat-settings-context'
 
-interface Props {
-  autoModel: boolean
-  onClick: () => void
-}
+export function AutoModelToggle() {
+  const { autoModel, handleToggleAutoModel } = useChatSettingsContext()
 
-export function AutoModelToggle({ autoModel, onClick }: Props) {
   return (
     <ToggleInput
       label='Auto Fetch Models'
       labelClassName='text-sm font-medium text-gray-700 dark:text-gray-300'
       value={autoModel}
-      onClick={onClick}
+      onClick={handleToggleAutoModel}
     />
   )
 }
